feat(collaborator-access): retire partner presets, add hotlink auto-submit
